chore: foundation api client
