+++ packages/client-app/src/components/AddressBar/AddressBar.vue
@@ -144,7 +144,7 @@ const updateSelectedServer = (server: ScalarListboxOption) => {
 <template>
   <div
     v-if="activeRequest && activeExample"
-    class="min-h-header flex flex-row items-center"
+    class="flex flex-row items-center"
     :class="[themeClasses.topContainer]">
     <!-- <div class="text-c-2 flex w-80 flex-row items-center gap-1 p-4">
       <ScalarIcon
@@ -160,7 +160,7 @@ const updateSelectedServer = (server: ScalarListboxOption) => {
         v-model="selectedRequest">
         <div
           :class="[
-            'text-xxs bg-b-1 relative flex w-[720px] flex-1 flex-row items-stretch rounded border p-[3px]',
+            'text-xxs bg-b-1 relative flex w-full lg:min-w-[720px] order-last lg:order-none flex-1 flex-row items-stretch rounded border p-[3px]',
             { 'rounded-b-none': open },
             { 'border-transparent': open },
           ]">
@@ -180,11 +180,12 @@ const updateSelectedServer = (server: ScalarListboxOption) => {
               @change="updateRequestMethod" />
             <ScalarListbox
               v-if="serverOptions"
+              class="h-auto"
               :options="serverOptions"
               :value="activeCollection?.selectedServerUid"
               @update:modelValue="updateSelectedServer">
               <ScalarButton
-                class="font-code text-sm whitespace-nowrap"
+                class="relative h-auto min-h-[24.5px] shrink-0 gap-1.5 overflow-hidden px-1 py-0 font-code"
                 size="sm"
                 variant="outlined">
                 {{
